build(*): define externals and dedupe dependencies from final lib
