+++ package.json
@@ -6,8 +6,8 @@
   "module": "lib/index.js",
   "typings": "lib/index.d.ts",
   "scripts": {
-    "build:dev": "rm -rf lib && rollup --config rollup.config.js --environment NODE_ENV:development",
-    "build:prod": "rm -rf lib && rollup --config rollup.config.js --environment NODE_ENV:production",
+    "build:dev": "rm -rf lib && rollup --config rollup.config.mjs --environment NODE_ENV:development",
+    "build:prod": "rm -rf lib && rollup --config rollup.config.mjs --environment NODE_ENV:production",
     "build:react": "npm run cem && node ./scripts/makeReact.mjs",
     "build:readme": "gulp build-readme",
     "write:stories": "npm run cem && node ./scripts/generateStories.mjs && npm run concat:stories",
@@ -110,7 +110,7 @@
     "plop": "^3.1.2",
     "postcss": "^8.2.13",
     "prettier": "^2.8.7",
-    "rollup": "^2.46.0",
+    "rollup": "^3.28.1",
     "rollup-plugin-generate-package-json": "^3.2.0",
     "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-postcss-lit": "^1.0.1",
