+++ src/base/dropdown-element.ts
@@ -1,6 +1,7 @@
 import type { StrictModifiers } from "@popperjs/core";
 import * as Popper from "@popperjs/core";
-import Dropdown from "bootstrap/js/src/dropdown.js";
+import Dropdown from "bootstrap/js/src/dropdown";
+import type { Dropdown as BsDropdown } from "bootstrap";
 import { property, state } from "lit/decorators.js";
 import { Ref, createRef } from "lit/directives/ref.js";
 import { SgdsDropdownItem } from "../components/Dropdown/sgds-dropdown-item";
@@ -28,7 +29,7 @@ export class DropdownElement extends SgdsElement {
   /** @internal */
   myDropdown: Ref<HTMLElement> = createRef();
   /** @internal */
-  bsDropdown: Dropdown = null;
+  bsDropdown: BsDropdown = null;
 
   /** @internal Controls auto-flipping of menu */
   @property({ type: Boolean, state: true })
