+++ src/typings/index.d.ts
@@ -1,3 +1,3 @@
-declare module 'bootstrap/js/src/collapse';
-declare module 'bootstrap/js/src/dropdown';
-declare module 'bootstrap/js/src/tooltip';
+declare module "bootstrap/js/src/collapse";
+declare module "bootstrap/js/src/dropdown";
+declare module "bootstrap/js/src/tooltip";
